fix: 평균 별점 오류 수정 및 타입 변환

+++ src/main/java/com/example/tripKo/domain/place/dto/response/info/TouristSpotResponse.java
@@ -17,7 +17,7 @@ public class TouristSpotResponse {
 
   private Long id;
   private String name;
-  private Float averageScore;
+  private double averageScore;
   private String mainImage;
   private List<Content> contents;
   private String address;
